requirements.txt
tweepy
python-dotenv
selenium